requirements.txt
django==2.0
django-ckeditor
django-grappelli